feat(test): integrate GTest and configuration
The test folder will be built separately from the main executable of the
project. Besides, I also add vcpkg-configuration for better manifest
mode.

I change the VCPKG_TARGET_TRIPLET to x64-mingw-static to be able to
link the packages and compile them using mingw compiller. This is
required for gtest installation.