feat(game): add tetromino locking and game over
When the current tetromino detects collisions at its bottom edge,
instead of only undoing the previous movement, it also locks its
position and register that on the grid.

After locking tetromino position on the grid, the game will determine if
it is over or check for rows cleared to update score.