refactor(game): add runGameLoop() method
Instead of having the game loop inside main.cpp, I place that method
inside game.cpp and run it in main. This makes the structure of game
and main cleaner. main.cpp will plays its role as an entry for the
executable and game.cpp is in charge of handling the game logic,
including game loop.